modules/dagster-ibis-bigquery/pyproject.toml
[project]
name = "dagster-ibis-bigquery"
version = "0.1.0"
description = "Library for interacting with Ibis Tables within a bigquery backend within dagster."
readme = "README.md"
authors = [
    { name = "James Blackwell", email = "33688964+j-blackwell@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.10.6,<1.10.14",
    "dagster-gcp-pandas>=0.26.6",
    "ibis-framework[bigquery]>=10.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "dagster-ibis",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
dagster-ibis = { path = "../dagster-ibis", editable = true }
